requirements.txt
opensearch-py==3.0.0
load-dotenv==0.1.0